+++ assets/js/lipnamo-generate-items.js
@@ -12,6 +12,12 @@ jQuery(document).ready(function($){
             post_author = $('select[name="lipnamo_post_author"]').val(),
             post_status = $('select[name="lipnamo_post_status"]').val(),
             post_thumbnails = $('input[name="lipnamo_thumbnails"]').val(),
+            post_title_min = $('input[name="length_title_min"]').val(),
+            post_title_max = $('input[name="length_title_max"]').val(),
+            post_excerpt_min = $('input[name="length_excerpt_min"]').val(),
+            post_excerpt_max = $('input[name="length_excerpt_max"]').val(),
+            post_body_min = $('input[name="length_content_min"]').val(),
+            post_body_max = $('input[name="length_content_max"]').val(),
             $progressBar = $('.lipnamo-generate__progress-bar'),
             $wpbody = $("#wpbody");
 
@@ -38,6 +44,9 @@ jQuery(document).ready(function($){
                 post_author: post_author,
                 post_status: post_status,
                 post_thumbnails: post_thumbnails,
+                post_title_length: post_title_min + ',' + post_title_max,
+                post_excerpt_length: post_excerpt_min + ',' + post_excerpt_max,
+                post_body_length: post_body_min + ',' + post_body_max,
                 post_step: i
             },
             success: function(response){